services/training/requirements.txt
fastapi
mlflow
pydantic
uvicorn
torch
torchvision
opencv-python
albumentations
scikit-learn
numpy
pillow
pyyaml
tqdm
